# docs/swagger.yaml
basePath: /
definitions:
...
  delete.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  ukk-smkn2_features_role_create.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_role_create.Request:
    properties:
      nama:
        type: string
    type: object
  ukk-smkn2_features_role_create.Response:
    properties:
      id:
        type: string
      nama:
        type: string
    type: object
  ukk-smkn2_features_role_delete.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_role_delete.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  ukk-smkn2_features_role_getall.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_role_getall.Response:
    properties:
      id:
        type: string
      nama:
        type: string
    type: object
  ukk-smkn2_features_role_getbyid.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_role_getbyid.Response:
    properties:
      id:
        type: string
      nama:
        type: string
    type: object
  ukk-smkn2_features_role_update.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_role_update.Request:
    properties:
      nama:
        type: string
    type: object
  ukk-smkn2_features_role_update.Response:
    properties:
      id:
        type: string
      nama:
        type: string
    type: object
...
  ukk-smkn2_features_siswa_delete.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  ukk-smkn2_features_siswa_getbyid.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_siswa_getbyid.Response:
    properties:
      jurusan:
        type: string
      kelas:
        type: string
      nama:
        type: string
      nis:
        type: string
      siswa_id:
        type: string
      user_id:
        type: string
    type: object
  ukk-smkn2_features_siswa_update.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  ukk-smkn2_features_siswa_update.Request:
    properties:
      jurusan:
        type: string
      kelas:
        type: string
      nama:
        type: string
      nis:
        type: string
      user_id:
        type: string
    type: object
  ukk-smkn2_features_siswa_update.Response:
    properties:
      jurusan:
        type: string
      kelas:
        type: string
      nama:
        type: string
      nis:
        type: string
      siswa_id:
        type: string
      user_id:
        type: string
    type: object
host: 127.0.0.1:8080
info:
  contact: {}
  description: Restful API using go fiber
  title: your own API application
  version: "1.0"
paths:
...
  /api/role:
    get:
      description: Get list of all roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ukk-smkn2_features_role_getall.Response'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_getall.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get All Roles
      tags:
      - role
    post:
      consumes:
      - application/json
      description: Create a new role
      parameters:
      - description: Role create request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ukk-smkn2_features_role_create.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_create.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_create.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_create.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create Role
      tags:
      - role
  /api/role/{id}:
    delete:
      description: Delete a role by ID
      parameters:
      - description: Role ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_delete.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_delete.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_delete.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_delete.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Role
      tags:
      - role
    get:
      description: Get single role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_getbyid.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_getbyid.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_getbyid.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Role by ID
      tags:
      - role
    put:
      consumes:
      - application/json
      description: Memperbarui data role dengan ID tertentu
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body data role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ukk-smkn2_features_role_update.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_update.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_update.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_update.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_role_update.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Role by ID
      tags:
      - role
...
  /api/siswa/{id}:
    delete:
      description: Menghapus data siswa berdasarkan ID
      parameters:
      - description: Siswa ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/delete.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_siswa_delete.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_siswa_delete.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_siswa_delete.ErrorResponse'
      summary: Delete Siswa
      tags:
      - siswa
    get:
      description: Mengambil data siswa berdasarkan ID
      parameters:
      - description: Siswa ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_siswa_getbyid.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_siswa_getbyid.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_siswa_getbyid.ErrorResponse'
      summary: Get Siswa by ID
      tags:
      - siswa
    put:
      consumes:
      - application/json
      description: Memperbarui data siswa dengan ID tertentu
      parameters:
      - description: Siswa ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body data siswa
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ukk-smkn2_features_siswa_update.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ukk-smkn2_features_siswa_update.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ukk-smkn2_features_siswa_update.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ukk-smkn2_features_siswa_update.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ukk-smkn2_features_siswa_update.ErrorResponse'
      summary: Update Siswa by ID
      tags:
      - siswa
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"